anacron: fix types in comparisons
This fixes warnings reported by -Wsign-compare compiler option.

+++ src/env.c
@@ -23,6 +23,7 @@
 
 #include <ctype.h>
 #include <errno.h>
+#include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
 #include <sys/types.h>
@@ -281,7 +282,7 @@ char *env_get(const char *name, char **envp) {
 	while ((p = *envp++) != NULL) {
 		if (!(q = strchr(p, '=')))
 			continue;
-		if ((q - p) == len && !strncmp(p, name, len))
+		if ((size_t)(q - p) == len && !strncmp(p, name, len))
 			return (q + 1);
 	}
 	return (NULL);
